jobs/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
bs4==0.0.2
cassandra-driver==3.29.2
certifi==2025.6.15
click==8.2.1
distro==1.9.0
geomet==0.2.1.post1
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
kafka-python==2.2.14
openai==1.93.0
playwright==1.53.0
py4j==0.10.9.9
pydantic==2.11.7
pydantic_core==2.33.2
pyee==13.0.0
pyspark==4.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
